Align marketing page headers
